Add platform identifier to slice, re-factor installer code
A slice is now specific to a kernel as well as a platform.  The
installers support the full cross-product of kernels and platforms.

The list of supported platforms is not yet complete, as is the
configuration of platform-dependent features of the control- and
data-plane.

The primary platform identifiers are those used as the "onie_machine"
ID at the ONIE level. The ONIE installer copies /etc/machine.conf to
the root file system, which makes the platform identifier available to
the RARE services.