Add support for the stordis_bf2556x_1t
This platform uses a gear box to support 1G interfaces. It uses a
"SAL" wrapper that launches bf_switchd to be able to modify the Tofino
ports connected to the gear box.  The bf_forwarder needs to talk to
both, bf_switchd and the wrapper via gRPC.  The wrapper uses its own
protobuf-based protocol. The python bindings for that API are provided
by the bf-platforms package for the APS BF2556 baseboard.