Use new jobset structure
This change re-organizes how various components of a release are
getting built. release.nix now has three independent switches to
enable building of

   * release and releaseClosure
   * standaloneInstaller
   * onieInstaller

The declarative jobset description in spec.json still contains one
jobset for the development version and one jobset per release. The
development jobset only builds release/releaseClosure as before.  The
release-specific jobsets no longer build the ONIE installer, but still
include the standaloneInstaller.

The ONIE installers can be created by a new jobset declaration that
sets the buildOnieInstaller argument of release.nix to true.  One
reason for this change is to eliminate the check for a "principal"
release in release.nix.  The other reason is to pass a list of
specific binary caches to include in the ONIE install image. These
caches are populated by the Hydra instance that produces the
installer. This mechanism binds all systems set up by the ONIE
installer to the same set of binary caches.