Add systemd services for bf_forwarder.py and bf_switchd, refactor p4 profile support
Run the bf_forwarder.py and bf_switchd processes as systemd services
rather than as subprocesses of freerouter. This facilitates service
management and adds logging in the system journal.

The P4 profile is now selected by writing the profile name to
/etc/freertr/p4-profile and restarting the freerouter service.

If needed, bf_forwarder.py can be run with profile-dependent arguments
by declaring a set of flags in rare/profiles.nix. In this case, a
shell wrapper around bf_forwarder is created that passes the dedicated
flags to the script.