Check for diffs before removing rtr-hw.txt
It is not yet clear whether rtr-hw.txt should really be immutable. We
want to allow the operator to replace the symlink with an editable
copy of the file.  During deactivation, it is checked wheter the file
differs from the original. If so, it is moved out of the way rather
than deleted. The new name is the old name with the first 6 digits of
the md5sum of the file appended.