Move to SDE 9.7.1
This release addresses an issue where a Tofino1 chip can stop
forwarding packets under specific circumstances.  A new compiler
option was introduced to safeguard against this, see issue P4C-3974 in
the 9.7.1 release notes.  The "bier" profile must disable this feature
to compile due to it's large headers and is therefore at risk to
trigger the problem.