Don't use "nohid" in rtr-hw.txt
The problem with the "nohid" feature is that it causes (among other
things) the "exec" line of the process to be added to the running
config. That line contains a path to the Nix store where the
bffwd/bfswd executables reside, which will break the release-manager
upgrade mechanism. To make use of nohid, we would have to use paths to
/nix/var/nix/profiles/RARE instead.

In general, it is questionable whether it is a good thing to modify
the running config when the rtr-hw.txt file changes after an upgrade.