Add aliases to invoke the release manager from the freeRtr CLI
Most functions of the release manager are now accesible from the
freeRtr CLI, incuding all available mechanisms to install versions.

The switch to a new version (RARE Nix profile generation) is scheduled
by the tna-switch-to-generation alias. It writes the generation to
switch to into a file in /etc/freertr.  The actual switch is performed
either by the ExecStopPost or ExecStartPre hooks of the freerouter
service unit, depending on whether the service is restarted by "reload
warm" or a restart of the service itself.

The P4 profile can be queried or set by the tna-set-profile alias.
The new profile is activated either by "reload warm" or a service
restart. Currently, "reload process bfswd" does not work because
freeRtr does not kill child processes of external processes. Once this
is fixed, a new profile can be activated by reloading the bfswd and
bffwd processes, which can be integrated into the tna-set-profile
alias.