Add ability to restrict platforms and kernels in installers
default.nix takes two more options installerPlatforms and
installerKernels to restrict what is included in the ONIE and
standalone installers. This is useful to create installers tailored
for a specific environment either to save a bit of space or to exclude
platforms for which the BSP is not available.

A Makefile is added that makes use of these options to build
customized installers from the Git repository.  The options can also
be used from Hydra jobsets to automate building of such installers.

The Makefile also contains a target "install" to perform a local
installation directly from the repository. This is done by first
building the release manager, then calling it with the option
`--install-local .` (this option was added in commit f8a8ca3 of
bf-sde-nixpkgs).