default.nix: add argument freerouterSrc
This argument allows the caller to supply a different version of the
freerouter repo to override what's specified in
overlay/freerouter/repo.nix. This will be used by a new version of the
release-manager to support installation of an "experimental" version
of RARE based on the newest freerouter code in a Nix-friendly manner.