Add --install-experimental option to the release-manager
This option installs the newest commit from the freerouter Git repo on
top of the newest commit from this repo in a Nix-friendly manner. The
command creates an ad-hoc Git tag by adding the string "-freertr-"
followed by the first 6 characters of the freerouter commit to the Git
tag of the tip of the master branch of the rare-nix repository. This
should provide a unique Id to identify the version in the RARE
profile.

This change requires changes in the bf-sde-nixpkgs support functions
provided by 0aa3b. This commit also changes the ad-hoc Git tag for the
--install-local mechanism by adding the string "-local-". This change
is mirrored in the Makefile's "standaloneInstaller" and
"onieInstaller" targets.

Unrelated to this, the onieInstaller Makefile target now also includes
the standard binary cache p4.cache.nix.net.switch.ch.