Bump bf-sde-nixpkgs version to 1710af3
This fixes an issue where building a module wrapper for an unsupported
kernel threw an error which broke certain version upgrades with the
release-manager.  The fix reverts back to the stragey used previously
where we produce a wrapper that terminates with an error rather than
throwing an error during Nix evaluation.