Make sure release-manager sees $NIX_SSL_CERT_FILE
NIX_SSL_CERT_FILE is only set for login shells. In particular, it is
not set when freertr is started from systemd. Hence, when
release-manager is called via an alias from freertr, it doesn't see it
either and the git API calls via HTTPS could fail due to missing
certificates.