Add a utility function to show all sets of compiler options for a profile
Every profile needs to be compiled for all supported platforms when
building a release.  In general, this leads to an array of sets of
distinct compiler options and the profile needs to be compiled for
each of them.  A platform implies a specific ASIC which fixes the
Tofino version (T1, T2, T3 etc.) and the number of pipes.

The RARE table optimizer must make sure that an optimized profile
builds successfully for all sets of compiler options. The new utlity
function flagsForProfile in the top-level expression takes the name of
a profile and a Tofino version as input (tofino, tofino2 etc.) and
returns the corresponding set of compiler flags for all supported
platforms of the given Tofino version. Usage example:

$ nix eval --json '(with import ./. {}; flagsForProfile "PE" "tofino")' | jq -r '(.[] | join(" "))'
-DPROFILE_PE -DQUAD_PIPE
-DDUAL_PIPE -DPROFILE_PE
-DPROFILE_PE