Update freertr and rare repos
This change enables passing the port speed configuration as a string
from freertr to bf-forwarder to support explicit configuration of the
number of serdes lanes where multiple choices are possible. The number
of lanes can be appended to the speed separated by a colon,
e.g. "100:2" selects 100G on two lanes. The available choices depend
on the ASIC type (Tofino, Tofino2). This parameter is optional.