Update bf-sde-nixpkgs to ecd155
   * Add SDE 9.13.4
   * Add Debian 12.6, 12.7, 12.8 kernels
   * Update onie-debian-nix-installer to a93b72

The new version of mkOnieInstaller supports adding an instance of the
SDE development shell to the ONIE installer. The feature can be
activated by seting the new withSdeEnv argument of the main Nix
expression to true.

The new option activate can be used to control the activation of the
service in a newly installed system. If set to true (default), the
service is active after the first boot of the system. If set to false,
the service must be activated explicitly when needed.